Add license, build scripts and documentation
Big changes. The main functions of the program were moved into a
separate module. Files were put in directory structures for use with
Module::Build. Build scripts were added, using the Module::Build. The
GPLv3 was put into the repository in the COPYING file, as well as
license notices in each source file. POD documentation was added to the
main script which gets installed on to the system as a man page.